refactor: split rule/matchers.rs into a matchers/ module directory
Part 2 of #1861 (port of hayabusa-pro#182). The ~4900-line
src/detections/rule/matchers.rs is broken into a matchers/ directory (mod.rs style,
matching src/detections/rule/mod.rs):

- matchers/mod.rs           — the LeafMatcher trait, module decls, public re-exports.
- matchers/leaf_matchers.rs — MinlengthMatcher, RegexesFileMatcher,
                              AllowlistFileMatcher.
- matchers/default_matcher.rs — DefaultMatcher and its MatchPlan/Wrap/Encoding/
                              Utf16Kind fast-path types.
- matchers/pipe_element.rs  — the PipeElement enum + its impl.
- matchers/tests.rs         — the unit tests.

Pure code movement — the only changes are relocated `use` imports, `mod`/`pub use`
declarations, and `pub(super)` visibility bumps; every function body is unchanged.
A normalized multiset diff against the original shows the only content differences
are two signatures rustfmt reflowed onto multiple lines (the `pub(super)` prefix
pushed them past 100 cols). Detection output is byte-identical to the pre-split
build on the full sample-evtx corpus (csv and json-timeline), the full lib suite
(483) passes, and fmt + clippy --all-targets -- -D warnings are clean.

The per-modifier trait / category reorganization from #1861 is left as a follow-up.
